moving swagger configuration to a new class
