+++ src/main/java/com/redhat/cajun/navy/responder/Application.java
@@ -2,25 +2,12 @@ package com.redhat.cajun.navy.responder;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
+
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
 @SpringBootApplication
-@EnableSwagger2
 public class Application {
-
-	@Bean
-    public Docket swagger() {
-        return new Docket(DocumentationType.SWAGGER_2)
-                .select()
-                .apis(RequestHandlerSelectors.basePackage("com.redhat.cajun.navy.responder"))
-                .paths(PathSelectors.any())
-                .build();
-    }
+    
 	public static void main(String[] args) {
 		SpringApplication.run(Application.class, args);
 	}
