+++ src/main/java/com/redhat/cajun/navy/responder/SwaggerConfiguration.java
@@ -1,0 +1,40 @@
+package com.redhat.cajun.navy.responder;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import springfox.documentation.builders.ApiInfoBuilder;
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.service.ApiInfo;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+
+@Configuration
+@EnableSwagger2
+public class SwaggerConfiguration {
+
+	@Bean
+    public Docket swagger() {
+        return new Docket(DocumentationType.SWAGGER_2)
+                .select()
+                .apis(RequestHandlerSelectors.basePackage("com.redhat.cajun.navy.responder"))
+                .paths(PathSelectors.any())
+                .build()
+                .apiInfo(apiInfo());
+    }
+	
+    private ApiInfo apiInfo() {
+    	 
+        ApiInfo apiInfo = new ApiInfoBuilder()
+                .title ("API Responder Service")
+                .description ("Emergency Response - Responder Service API")
+                .version("1.0.0")
+                .build();
+ 
+        return apiInfo;
+    }
+
+    
+}
